+++ StudiECS/ChunkTest.h
@@ -11,7 +11,7 @@ void ChunkTest()
     std::array<int, 18> int_array;
     std::fill(int_array.begin(), int_array.end(), -1);
 
-    ArrayView<int> int_view(int_array.data(), int_array.size());
+    ArrayView<int> int_view(int_array.data(), static_cast<uint32>(int_array.size()));
     int_view[1] = 56;
     int_view[3] = 35;
     int_view[8] = 7;
@@ -24,7 +24,7 @@ void ChunkTest()
     std::array<Bool, 1024> bool_array;
     std::fill(bool_array.begin(), bool_array.end(), false);
 
-    ArrayView<bool> bool_view(bool_array.data(), bool_array.size());
+    ArrayView<bool> bool_view(bool_array.data(), static_cast<uint32>(bool_array.size()));
     bool_view[0] = true;
     bool_view[1023] = true;
     bool_view[3] = true;
